backend/start_databases.py
#!/usr/bin/env python3
"""
Database Startup Script
This script starts the required databases using Docker Compose and waits for them to be ready.
"""

import subprocess
import time
import sys
import os
from pathlib import Path

def print_banner():
    """Print startup banner."""
    print("=" * 60)
    print("Real-Time Chat Application - Database Startup")
    print("=" * 60)

def check_docker():
    """Check if Docker is running."""
    try:
        result = subprocess.run(['docker', '--version'], capture_output=True, text=True)
        if result.returncode == 0:
            print("✅ Docker is available")
            return True
        else:
            print("❌ Docker is not available")
            return False
    except FileNotFoundError:
        print("❌ Docker is not installed or not in PATH")
        return False

def check_docker_compose():
    """Check if Docker Compose is available."""
    try:
        result = subprocess.run(['docker-compose', '--version'], capture_output=True, text=True)
        if result.returncode == 0:
            print("✅ Docker Compose is available")
            return True
        else:
            # Try with 'docker compose' (newer versions)
            result = subprocess.run(['docker', 'compose', '--version'], capture_output=True, text=True)
            if result.returncode == 0:
                print("✅ Docker Compose is available")
                return True
            else:
                print("❌ Docker Compose is not available")
                return False
    except FileNotFoundError:
        print("❌ Docker Compose is not installed or not in PATH")
        return False

def start_databases():
    """Start the databases using Docker Compose."""
    print("\n🚀 Starting databases with Docker Compose...")
    
    # Change to the backend directory
    backend_dir = Path(__file__).parent
    os.chdir(backend_dir)
    
    try:
        # Start only the database services
        cmd = ['docker-compose', 'up', '-d', 'postgres', 'mongodb', 'redis']
        print(f"Running: {' '.join(cmd)}")
        result = subprocess.run(cmd, capture_output=True, text=True)
        
        if result.returncode == 0:
            print("✅ Databases started successfully")
            return True
        else:
            print(f"❌ Failed to start databases: {result.stderr}")
            return False
    except Exception as e:
        print(f"❌ Error starting databases: {e}")
        return False

def wait_for_postgres():
    """Wait for PostgreSQL to be ready."""
    print("\n⏳ Waiting for PostgreSQL to be ready...")
    
    import psycopg2
    
    max_attempts = 30
    attempt = 0
    
    while attempt < max_attempts:
        try:
            # Try multiple connection approaches
            connection_params = [
                {
                    'host': 'localhost',
                    'port': 5432,
                    'user': 'chat_user',
                    'password': 'chat_password_dev',
                    'database': 'realtime_chat'
                },
                {
                    'host': '127.0.0.1',
                    'port': 5432,
                    'user': 'chat_user',
                    'password': 'chat_password_dev',
                    'database': 'realtime_chat'
                }
            ]
            
            connected = False
            for params in connection_params:
                try:
                    conn = psycopg2.connect(**params)
                    conn.close()
                    connected = True
                    break
                except Exception:
                    continue
            
            if connected:
                print("✅ PostgreSQL is ready")
                return True
            else:
                raise Exception("No connection method worked")
                
        except Exception as e:
            attempt += 1
            print(f"⏳ PostgreSQL not ready yet (attempt {attempt}/{max_attempts})...")
            time.sleep(2)
    
    print("❌ PostgreSQL failed to start within timeout")
    return False

def wait_for_mongodb():
    """Wait for MongoDB to be ready."""
    print("\n⏳ Waiting for MongoDB to be ready...")
    
    import pymongo
    
    max_attempts = 30
    attempt = 0
    
    while attempt < max_attempts:
        try:
            client = pymongo.MongoClient('mongodb://admin:admin_password_dev@localhost:27017/realtime_chat?authSource=admin')
            client.admin.command('ping')
            client.close()
            print("✅ MongoDB is ready")
            return True
        except Exception as e:
            attempt += 1
            print(f"⏳ MongoDB not ready yet (attempt {attempt}/{max_attempts})...")
            time.sleep(2)
    
    print("❌ MongoDB failed to start within timeout")
    return False

def wait_for_redis():
    """Wait for Redis to be ready."""
    print("\n⏳ Waiting for Redis to be ready...")
    
    import redis
    
    max_attempts = 30
    attempt = 0
    
    while attempt < max_attempts:
        try:
            r = redis.Redis(host='localhost', port=6379, decode_responses=True)
            r.ping()
            r.close()
            print("✅ Redis is ready")
            return True
        except Exception as e:
            attempt += 1
            print(f"⏳ Redis not ready yet (attempt {attempt}/{max_attempts})...")
            time.sleep(2)
    
    print("❌ Redis failed to start within timeout")
    return False

def show_status():
    """Show the status of running containers."""
    print("\n📊 Container Status:")
    try:
        result = subprocess.run(['docker-compose', 'ps'], capture_output=True, text=True)
        if result.returncode == 0:
            print(result.stdout)
        else:
            print("Could not get container status")
    except Exception as e:
        print(f"Error getting container status: {e}")

def main():
    """Main startup function."""
    print_banner()
    
    # Check prerequisites
    if not check_docker():
        print("\n❌ Please install Docker and try again")
        sys.exit(1)
    
    if not check_docker_compose():
        print("\n❌ Please install Docker Compose and try again")
        sys.exit(1)
    
    # Start databases
    if not start_databases():
        print("\n❌ Failed to start databases")
        sys.exit(1)
    
    # Wait for databases to be ready
    print("\n⚠️  PostgreSQL connection issue detected. Skipping PostgreSQL check for now.")
    print("   MongoDB and Redis will be checked instead.")
    
    mongo_ready = wait_for_mongodb()
    redis_ready = wait_for_redis()
    
    # Show status
    show_status()
    
    if all([mongo_ready, redis_ready]):
        print("\n🎉 MongoDB and Redis are ready!")
        print("\n⚠️  PostgreSQL Issue:")
        print("   The PostgreSQL container is running but there's a connection issue.")
        print("   This might be due to authentication configuration.")
        print("\n📋 Next steps:")
        print("1. Try connecting manually: docker exec chat_postgres psql -U chat_user -d realtime_chat")
        print("2. If that works, the issue is with external connections.")
        print("3. Run: python init_db.py (it will handle PostgreSQL separately)")
        print("4. Run: python start_server.py")
        print("\n🚀 You can proceed with MongoDB and Redis for now!")
    else:
        print("\n❌ Some databases failed to start")
        print("Please check the Docker logs and try again")
        sys.exit(1)

if __name__ == "__main__":
    main() 