backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Rate Limiting
slowapi==0.1.9

# Monitoring and Logging
structlog==23.2.0
prometheus-fastapi-instrumentator==6.1.0

# AI/ML
openai==1.3.7

# File Handling
python-magic==0.4.27
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# CORS
python-cors==1.7.0